drop erroneous TODO
In a weak sense, doing what the TODO says would be wrong in general,
because while the article at the end of the link may originate from
its author, the link in the Perl Weekly does not.

More importantly, however, it would be categorically incorrect for
any link with a short summary or comment written by a Perl Weekly
editor, because the implied claim in the metadata would be that that
summary or comment was written by the linked article’s author.

(This type of semantic confusion is why the Atom syndication format
has the `source` element.)